starting removing level from trail
Adding the level in the trail is only an optimisation: it is needed in
an implementation, but it is useless from a theoretic point of view.

The properties of the alternative form (i.e., counting the decided
literals right of the literals has different properties w.r.t to
duplicate literals).

The next step is fully remove the annotation from the ``Decided''
constructor.