tuning proofs
In this commit the Isar proof of the from "obtain E where E = ..." have
been removed (they date back to the previous version of the locale).

The habd-tuned induction principles _lev2 have been removed: the
assumption "undefined M1 L" is not needed thanks to the simplifications
of the assumptions of the cdcl_W locale.