raw_init_+ raw_learned_clss ~> raw_ + learned_clss
The aim of this change is to improve reusability by better
abstraction. Instead of having two 'clss in the 2-WL, we use only
one and have a criterion whteher a clause is learned or not. This also
means less operations on the 'clss

This will ease the move to iterators instead of objects that are
convertible to multiset like lists.