paramtrisation of ~ in CDCL_W
Instead of defining \<sim> as the equality on top of the states, it is a
parameter of the locale. This allows to add fields to the state (like
the weight or a number representing the number of steps) and
still ensures that the rule do *not* modifiy the aformentionned fields.