more; starting adding outlearned to express the minimisation;
Currently, the minimisation can only be expressed with the conflict
clause as lookup table (as a concrete algorithm) or in an abstract
version (via specification).

The first version should be avoided if possible (the representation is
not important) and the latter is only possible when the state also
containes the unit (and entailed) clauses.

Adding an "outlearned" component (as called in MiniSAT) allows to
express minisation without the iteration problem associated to the
lookup table. Therefore, this will make the proofs shorter (and the
solver might become slightly more efficient, especially since it
allows to change the lookup take from literal to atmo only).