qsort: invariant was wrong
I observed that the invariant of quicksort was not only overly
complicated, it was also wrong.  Actually, we don't have an invariant at
the first place; we have a precondition and a postcondition.  In this
commit, I wrote down new conditions.  The precondition just requires that
the pointers are valid.  The postcondition claims that the output is a
sorting in the given bounds, and the list was not changed otherwise.  I
outlined the proof and identified the cases, for which I admitted the
proofs.  I will write lemmas for each of the cases.