IsaSAT: use stamping for lbd instead of cleaning each time
Previously, an array of bools was used that had to be emptied each
time after marking another clause. We now use a stamp (a natural
number): The value is present iff we have stamp at the position.
Cleaning is very cheap most of the time (increment stamp) and
expensive before overflowing. This is an obvious and trivial
improvement.