isasat: more regular rephasing
Before rephasing was based on the number of clauses, which is a
strange choice. It is now based on the number of
conflicts. Additionally, we now rephase even if a restart was
done. This should make scheduling bugs easier to identify.

We also print the same letters as cadical instead of numbers.