forward: start to strengthen invariant
We attempted to allow for units in the candidates for forward
subsumption, but this turned to be too complicated. So we started to
strengthen the invariants, requiring various changing in the
strengthening algorithm. For example, we want to ensure that the
candidate clause is not in the occurence lists and has not been deleted.